constify the Eq, PartialOrd and Ord traits
Also constify the impls of basic types.

One potentially controversial part of this change is making Eq, a
marker trait, const. I chose to do this ease user adoption.
Otherwise, code which already has an Eq bound and uses it to
proxy a PartialEq bound would need to add a separate bound. This
would cause a litering of bounds in downstream types.